requirements.txt
Jinja2==2.10.1
PyYAML
termcolor